requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.15.1,<0.16
djangorestframework-simplejwt>=5.0.0,<5.3
pandas>=2.0.0


